# docs/swagger.yaml
definitions:
  model.LoginRequestType:
    properties:
      id:
        type: string
      password:
        type: string
    type: object
  model.RegisterRequestType:
    properties:
      birthday:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
  model.SignoutRequestType:
    properties:
      password:
        type: string
    type: object
  model.UpdateRequestType:
    properties:
      email:
        type: string
      id:
        type: string
      password:
        type: string
    type: object
  model.User:
    properties:
      birthday:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      user_uuid:
        type: string
    type: object
  utility.FailResponse:
    properties:
      code:
        type: integer
      error:
        type: string
      message:
        type: string
      status:
        type: string
    type: object
  utility.LoginResponse:
    properties:
      code:
        type: integer
      message:
        type: string
      status:
        type: string
      token:
        type: string
    type: object
  utility.OKResponse:
    properties:
      code:
        type: integer
      message:
        type: string
      status:
        type: string
    type: object
  utility.ProfileResponse:
    properties:
      code:
        type: integer
      message:
        type: string
      status:
        type: string
      user:
        $ref: '#/definitions/model.User'
    type: object
info:
  contact: {}
paths:
  /:
    get:
      consumes:
      - application/json
      description: 책 목록을 query와 target을 통해 불러옵니다. target은 title, isbn, publisher,
        person이 있으며, query는 검색어입니다.
      parameters:
      - description: query를 통해 검색어 입력
        in: query
        name: query
        required: true
        type: string
      - description: target을 통해 검색 타입 지정
        in: query
        name: target
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utility.OKResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utility.FailResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utility.FailResponse'
      summary: 책목록을 조회합니다.
      tags:
      - book
  /login:
    post:
      consumes:
      - application/json
      description: 로그인 후 토큰을 반환합니다.
      parameters:
      - description: 로그인을 위해 필요한 유저 아이디 및 암호화 되지 않은 비밀번호
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.LoginRequestType'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utility.LoginResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utility.FailResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utility.FailResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utility.FailResponse'
      summary: 로그인 기능입니다.
      tags:
      - user
  /user:
    delete:
      consumes:
      - application/json
      description: 유저 정보를 제거 합니다, 토큰과 비밀번호를 입력받습니다.
      parameters:
      - description: token을 통해 유저 조회
        in: query
        name: token
        required: true
        type: string
      - description: 비밀번호를 통해 본인 확인
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.SignoutRequestType'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utility.OKResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utility.FailResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utility.FailResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utility.FailResponse'
      summary: 유저 정보를 데이터베이스에서 제거합니다.
      tags:
      - user
    get:
      consumes:
      - application/json
      description: 유저의 프로필 정보를 token을 통해 조회합니다.
      parameters:
      - description: token을 통해 유저 조회
        in: query
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utility.ProfileResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utility.FailResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utility.FailResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utility.FailResponse'
      summary: 유저의 프로필을 조회합니다
      tags:
      - user
    post:
      consumes:
      - application/json
      description: 회원가입을 통해 데이터베이스에 유저 객체를 삽입합니다.
      parameters:
      - description: 회원가입을 위해 필요한 유저 아이디, 암호화 되지 않은 비밀번호, 이메일, 이름 및 생일
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.RegisterRequestType'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/utility.OKResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utility.FailResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/utility.FailResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utility.FailResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utility.FailResponse'
      summary: 회원가입 기능입니다.
      tags:
      - user
  /user/check/validate:
    post:
      consumes:
      - application/json
      description: id 또는 email중 하나만 검사하시길 바랍니다.
      parameters:
      - description: id 또는 email중 하나
        in: query
        name: filter
        required: true
        type: string
      - description: 중복 검사할 데이터
        in: query
        name: data
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utility.OKResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/utility.FailResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utility.FailResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utility.FailResponse'
      summary: id, email의 중복 검사를 합니다.
      tags:
      - user
  /user/forgot/password:
    patch:
      consumes:
      - application/json
      description: 비밀번호를 잊었을 경우 변경합니다.
      parameters:
      - description: 비밀번호를 통해 본인 확인
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.UpdateRequestType'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utility.OKResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utility.FailResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utility.FailResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utility.FailResponse'
      summary: 비밀번호를 변경합니다.
      tags:
      - user
swagger: "2.0"